requirements.txt
PyQt6>=6.6
APScheduler>=3.10
watchdog>=4.0
feedparser>=6.0
httpx>=0.27
pyyaml>=6.0
pydantic>=2.6
beautifulsoup4>=4.12
lxml>=5.0
# Transitive — pinned explicitly so py2app bundles them
sniffio>=1.3
anyio>=4.0
h11>=0.14
httpcore>=1.0
pytz>=2024.1
tzlocal>=5.0
# Security: hardened XML parser used in OPML import (defuses XXE + bombs)
defusedxml>=0.7
